Move providers config to a new file

# Conflicts:
#	packages/strapi-plugin-users-permissions/services/Providers.js
